+++ Assets/SS_WorkSpace/Scripts/RandomSelect.cs
@@ -7,7 +7,7 @@ public class RandomSelect : MonoBehaviour
     public List<GameObject> towers = new List<GameObject>();
     public Transform spawnPosition;
     
-    int total = 0;
+    float total = 0;
     GameObject result;
     
     public void ResultSelect()
@@ -18,7 +18,7 @@ public class RandomSelect : MonoBehaviour
     
     public GameObject RandomTower() //호출시 타워 리스트에서 가중치를 통한 임의의 타워 반환
     {
-        int weight = 0;                                                   //가중치 변수
+        float weight = 0;                                                   //가중치 변수
         int selectNum = Mathf.RoundToInt(total*Random.Range(0.0f,1.0f)); //실수 0~1사이의 임의의값을 total에 곱함
         
         for (int i = 0; i < towers.Count; i++)
